# Faster histograms