--- a/mypy.ini
+++ b/mypy.ini
@@ -13,3 +13,6 @@ ignore_missing_imports = True
 
 [mypy-torchkbnufft.*]
 ignore_missing_imports = True
+
+[mypy-fast_histogram.*]
+ignore_missing_imports = True