--- a/src/mpol/gridding.py
+++ b/src/mpol/gridding.py
@@ -246,12 +246,14 @@ class GridderBase:
             A 2D array of size ``(npix, npix)`` in ground format containing the summed cell quantities.
         """
 
-        bin_range = self.coords.vis_ext[:2]
         result = fast_hist.histogram2d(
             vv,
             uu,
             bins=self.coords.ncell_u,
-            range=[bin_range, bin_range],
+            range=[
+                [self.coords.v_bin_min, self.coords.v_bin_max],
+                [self.coords.u_bin_min, self.coords.u_bin_max],
+            ],
             weights=values,
         )