I benchmarked a small part of the code base and found that the `np.histogram2d` call in `_sum_cell_values_channel` method is pretty slow. The method takes up almost 75% of the entire running time (22.2 s / 29.6 s). See the visualized profiling below with the relevant section highlighted in magenta.
<img width="1258" alt="image" src="https://user-images.githubusercontent.com/23248101/222869358-463ead5a-c4c8-4bd4-838b-6f0ed8eba755.png">
Since our gridding process creates grids that are uniformly spaced, we can opt for libraries that specialize in histograms for uniform bin sizes. I added `fast-histogram` into project's dependency and substituted the histogram call. The method call now only takes 6.2% of the total running time (0.565 s / 9.12 s) resulting in a 3.25x speed up. Relevant section highlighted again in magenta below.
<img width="1258" alt="image" src="https://user-images.githubusercontent.com/23248101/222869775-fb14f718-aa00-40d6-b964-7009263f7aa5.png">