@kadri-nizam 2023-03-04
> otherwise I would be cautious about adding another core dependency outside of numpy.

I concur. I'll look into whether numpy has routines that are made specially for uniform bin sizes.