@kadri-nizam 2023-03-05
It looks like the author of `fast-histogram` also contributed to making [numpy's histogram faster](https://github.com/numpy/numpy/pull/6100). They made their own version that's fully implemented in C without the bells and whistles that numpy does.