@iancze 2023-03-05
Looks great, thanks! Given the fast-histogram author's commitment to various open source projects, I don't think it's a huge risk that that package will become outdated before MPoL does :) And, if it does, it should be easy enough to switch back to numpy-only.